requirements.txt
google-generativeai>=0.8.0
requests>=2.31.0
tqdm>=4.66.0
torch>=2.0.0
setuptools>=69.0.0
